reddit_crawler/requirements.txt
requests
kafka-python
torch --index-url https://download.pytorch.org/whl/cpu
transformers
